requirements.txt
# =========================
# Web Framework
# =========================
Flask==3.0.0
Werkzeug==3.0.1

# =========================
# Environment Variables
# =========================
python-dotenv==1.0.1

# =========================
# HTTP Requests & OSINT
# =========================
requests==2.31.0
urllib3==2.1.0
beautifulsoup4==4.12.3

# =========================
# Data Processing
# =========================
pandas

# =========================
# Google Gemini AI
# =========================
google-generativeai==0.4.1

# =========================
# JSON & Utilities
# =========================
simplejson==3.19.2

# =========================
# Security & Validation
# =========================
email-validator==2.1.1

# =========================
# Optional (Stability & Debugging)
# =========================
gunicorn==21.2.0
